# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GaskeunKerja for Business \u2014 MVP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/candidates": {
      "post": {
        "tags": [
          "candidates"
        ],
        "summary": "Create Candidate",
        "description": "HR/seed-side only for MVP \u2014 not a public candidate-facing endpoint (see Area 1 T8 note).\n\nRound-3 follow-up #8 (2026-07-19): CV upload now runs BOTH pipelines synchronously \u2014 CV\nextraction (ingest_cv, unchanged) AND skill-gap analysis + match scoring (compute_match_score,\nnewly wired in here). Previously compute_match_score() was only ever called from one-off seed\nscripts (load_demo_data.py/load_t5_fixture.py), never from this live endpoint \u2014 a real\ncandidate uploaded through the running app never got a match_scores row at all, and so never\nappeared on the Shortlist page (get_ranked_candidates only lists candidates that already have\none). This makes upload the primary path for scoring; the self-heal fallback in\nget_or_compute_skill_gap() remains as a defensive backstop for edge cases (e.g. a JD's\ncompetencies changing after upload), not the primary mechanism going forward.\n\nSynchronous by explicit choice over a background task: adds ~25-30s (5 LLM calls) on top of\nCV-parsing's existing latency, but keeps this endpoint fully consistent with ingest_cv's\nalready-synchronous style \u2014 no new background-task/error-reporting infra needed this late in\nthe build.",
        "operationId": "create_candidate_candidates_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_candidate_candidates_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/candidates/{candidate_id}/questions": {
      "get": {
        "tags": [
          "interview_answers"
        ],
        "summary": "List Candidate Questions",
        "description": "Candidate-facing (token-authenticated). Only ever returns approved questions \u2014\ncandidates never see drafts. Each question now carries its own HR-set time limit\n(2026-07-19 follow-up, replacing the earlier job-wide-only duration) \u2014 the video recorder\nreads duration_seconds off the current question, not a single job-level value.",
        "operationId": "list_candidate_questions_candidates__candidate_id__questions_get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateQuestionOut"
                  },
                  "title": "Response List Candidate Questions Candidates  Candidate Id  Questions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/candidates/{candidate_id}/answers": {
      "post": {
        "tags": [
          "interview_answers"
        ],
        "summary": "Submit Interview Answer",
        "description": "Candidate-facing (token-authenticated, not HR JWT). Consent-gated: 403 without a\nconsent_records row for this candidate.\n\nRound-3 follow-up #22 (2026-07-19): only saves the video synchronously now \u2014 transcription +\nrubric scoring (the slow part, 10-20+ seconds per answer) run afterward as a BackgroundTask,\nso this response returns as soon as the file is on disk. User-reported problem this fixes:\nwith all 4 answers now uploaded sequentially at the end of the interview (T17-followup #21),\nthe candidate was stuck watching \"Mengunggah jawaban N dari 4...\" for the FULL synchronous\npipeline time of every single answer, back to back.\n\n2026-07-22 (user decision, reversing an earlier attempt at a blocking \"please wait\" screen):\nthe candidate is released the moment their video data is confirmed stored \u2014 not once the\nLLM chain (transcription/scoring/summary/upskilling plan) finishes. That pipeline now runs\nfully invisibly in the background; only HR-facing surfaces (Laporan list, report page) gate on\nit actually completing (routers/jobs.py's processing_complete, routers/report.py's\n_require_report_ready).",
        "operationId": "submit_interview_answer_candidates__candidate_id__answers_post",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_interview_answer_candidates__candidate_id__answers_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnswerOut": {
        "properties": {
          "answer_id": {
            "type": "integer",
            "title": "Answer Id"
          },
          "audio_path": {
            "type": "string",
            "title": "Audio Path"
          }
        },
        "type": "object",
        "required": [
          "answer_id",
          "audio_path"
        ],
        "title": "AnswerOut"
      },
...
      "Body_create_candidate_candidates_post": {
        "properties": {
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "alias": {
            "type": "string",
            "title": "Alias"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "alias",
          "file"
        ],
        "title": "Body_create_candidate_candidates_post"
      },
      "Body_submit_interview_answer_candidates__candidate_id__answers_post": {
        "properties": {
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "session": {
            "type": "string",
            "title": "Session"
          },
          "token": {
            "type": "string",
            "title": "Token"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "session",
          "token",
          "file"
        ],
        "title": "Body_submit_interview_answer_candidates__candidate_id__answers_post"
      },
...
      "CandidateOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "alias": {
            "type": "string",
            "title": "Alias"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "alias"
        ],
        "title": "CandidateOut"
      },
      "CandidateQuestionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_text",
          "order_index",
          "duration_seconds"
        ],
        "title": "CandidateQuestionOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}